requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# Celery
celery==5.3.4

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1

# Utils
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Logging
structlog==23.2.0 